+++ Additionals/ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Recognition service added for automatic detection of sample values
 * Request is shown if a sample with value shall be removed
+* Number of periods is saved in settings
 * Menu entries reordered
 
 ## Version 1.3.0